mana hud 50% ajustado

+++ src/main/java/com/github/uranus_mod_group/uranus_mod/client/ManaHudOverlay.java
@@ -60,7 +60,7 @@ public class ManaHudOverlay {
                 tamBarX,tamBarY);
 
         //mana value bellow mana_bar
-        String textMana = Double.toString(ClientManaData.getPlayerMana());
+        String textMana = Integer.toString((int)ClientManaData.getPlayerMana());
         GuiComponent.drawString(poseStack, minecraft.font, textMana, x - ( tamBarX/2 + 30), y - (tamBarY + 13), 0x1B82AB);
 
         //structure (details)
